+++ background/redux-slices/accounts.ts
@@ -60,8 +60,6 @@ export type CompleteAssetAmount = AnyAssetAmount &
   AssetMainCurrencyAmount &
   AssetDecimalAmount
 
-const USER_VALUE_DUST_THRESHOLD = 2
-
 // Comparator for two transactions by block height. Can be used to sort in
 // descending order of block height, with unspecified block heights (i.e.,
 // unconfirmed transactions) at the front of the list in stable order.
@@ -359,113 +357,5 @@ export const addAddressNetwork = createBackgroundAsyncThunk(
   async (addressNetwork: AddressNetwork, { dispatch }) => {
     dispatch(loadAccount(addressNetwork.address))
     await emitter.emit("addAccount", addressNetwork)
-  }
-)
-
-export const getAccountState = (state: {
-  account: AccountState
-}): AccountState => state.account
-
-export const getAssetsState = (state: { assets: AssetsState }): AssetsState =>
-  state.assets
-
-// FIXME This should probably live somewhere else.
-type FullState = {
-  account: AccountState
-  assets: AssetsState
-  ui: UIState
-}
-
-// FIXME This should probably live somewhere else.
-export const getFullState = (
-  state: FullState
-): { account: AccountState; assets: AssetsState; ui: UIState } => state
-
-export const selectAccountAndTimestampedActivities = createSelector(
-  getAccountState,
-  getAssetsState,
-  selectHideDust,
-  (account, assets, hideDust) => {
-    // TODO What actual precision do we want here? Probably more than 2
-    // TODO decimals? Maybe it's configurable?
-    const desiredDecimals = 2
-    // TODO Make this a setting.
-    const mainCurrencySymbol = "USD"
-
-    // Derive activities with timestamps included
-    const activity = account.combinedData.activity.map((activityItem) => {
-      const isSent =
-        activityItem.from.toLowerCase() ===
-        Object.keys(account.accountsData)[0].toLowerCase()
-
-      return {
-        ...activityItem,
-        ...(activityItem.blockHeight && {
-          timestamp: account?.blocks[activityItem.blockHeight]?.timestamp,
-        }),
-        isSent,
-      }
-    })
-
-    // Keep a tally of the total user value; undefined if no main currency data
-    // is available.
-    let totalMainCurrencyAmount: number | undefined
-
-    // Derive account "assets"/assetAmount which include USD values using
-    // data from the assets slice
-    const accountAssets = account.combinedData.assets
-      .map<CompleteAssetAmount>((assetItem) => {
-        const assetPricePoint = selectAssetPricePoint(
-          assets,
-          assetItem.asset.symbol,
-          mainCurrencySymbol
-        )
-
-        if (assetPricePoint) {
-          const enrichedAssetAmount = enrichAssetAmountWithDecimalValues(
-            enrichAssetAmountWithMainCurrencyValues(
-              assetItem,
-              assetPricePoint,
-              desiredDecimals
-            ),
-            desiredDecimals
-          )
-
-          if (typeof enrichedAssetAmount.mainCurrencyAmount !== "undefined") {
-            totalMainCurrencyAmount ??= 0 // initialize if needed
-            totalMainCurrencyAmount += enrichedAssetAmount.mainCurrencyAmount
-          }
-
-          return enrichedAssetAmount
-        }
-
-        return enrichAssetAmountWithDecimalValues(assetItem, desiredDecimals)
-      })
-      .filter((assetItem) => {
-        const isNotDust =
-          typeof assetItem.mainCurrencyAmount === "undefined"
-            ? true
-            : assetItem.mainCurrencyAmount > USER_VALUE_DUST_THRESHOLD
-        const isPresent = assetItem.decimalAmount > 0
-
-        // Hide dust and missing amounts.
-        return hideDust ? isNotDust && isPresent : isPresent
-      })
-
-    return {
-      combinedData: {
-        assets: accountAssets,
-        totalMainCurrencyValue: totalMainCurrencyAmount
-          ? formatCurrencyAmount(
-              mainCurrencySymbol,
-              totalMainCurrencyAmount,
-              desiredDecimals
-            )
-          : undefined,
-        activity: account.combinedData.activity,
-      },
-      accountData: account.accountsData,
-      activity,
-    }
   }
 )
